SGP_Mitra-main/requirements.txt
Flask 
pymongo 
flask-jwt-extended 
flask-socketio
flask-mail
python-dotenv
gunicorn
flask_cors
bcrypt
authlib
certifi
requests
Flask 
pymongo 
flask-jwt-extended 
flask-socketio
flask-mail
python-dotenv
gunicorn
flask_cors
bcrypt
authlib
certifi
requests
uuid
transformers
langchain
langchain_community
openai
vaderSentiment
langchain_openai
pinecone
sentence_transformers
flask_login
requests_oauthlib
opencv-python
pandas
numpy
matplotlib
seaborn
tensorflow
tf-keras
fpdf
huggingface_hub
torch
sounddevice
openai-whisper
edge-tts
pydub
torchaudio
torchvision
soundfile
langchain_openai
openpyxl